step one to an easier reading routes folder
